pyproject.toml
[project]
name = "dbeaver-gcloud-free"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.14"
authors = [{ name = "Ruslan Khasanshin" }]
maintainers = [{ name = "Ruslan Khasanshin" }]

dependencies = ["pydantic>=2.12.5", "pyyaml>=6.0.3", "questionary>=2.1.1"]

[dependency-groups]
dev = [
  "bandit>=1.9.4",
  "prek>=0.3.8",
  "pysentry-rs>=0.4.3",
  "ruff>=0.15.9",
  "ty>=0.0.28",
]

[tool.uv]
link-mode = "copy"

[tool.ruff]
target-version = "py314"
line-length = 100
extend-exclude = [".venv"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "UP",
  "B",
  "SIM",
  "ARG",
  "PL",
  "RUF",
  "C4",
  "FA",
  "N",
  "PERF",
  "W",
]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true

[tool.pysentry.defaults]
fail_on = "low"

[tool.pysentry.sources]
enabled = ["pypa", "pypi", "osv"]
